% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add.SMI.Function.R
\name{add.SMI}
\alias{add.SMI}
\title{Add estimated Soil Moisture Index (SMI) to weather data}
\usage{
add.SMI(weather, verbose = TRUE, tif.dir = NULL)
}
\arguments{
\item{weather}{A list of length 2. Weather data as outputted from the \code{\link[=get.SILO.weather]{get.SILO.weather()}}. Weather covariate names for list items should include:
\itemize{
\item \code{daily_rain}
\item \code{max_temp}
}}

\item{verbose}{Logical. Should progress be printed? Default if TRUE.}

\item{tif.dir}{Directory path where SLGA TIFF files are saves from using the \code{\link[=dl.slga]{dl.slga()}} function. If left as the default (\code{NULL}), files
will be downloaded from the SLGA API which may be time consuming with slow internet download speeds.}
}
\value{
The same \code{weather} list object as inputted with the addition of a matrix of daily SMI values in \code{...$data$smi}.
}
\description{
This function uses daily rainfall and temperature to estimate a SMI from a weather data object that has been output from
get weather functions such as \code{\link[=get.SILO.weather]{get.SILO.weather()}}.
}
\details{
A simplified SMI is calculated based on a dynamic water balance model where soil moisture is proportionally dried by daily evapotranspiration
and increased by daily precipitation. The equation for SMI for a following day(\eqn{SMI_{i+1}}) is:
\deqn{SMI_{i+1} = \min{(1,\frac{SMI_i}{\frac{PET_i}{AWC \times c_s}c_d + 1} + (P_i \times c_p))}}
where \eqn{SMI_i} is the SMI for day \eqn{i}, \eqn{PET_i} is the potential evapotranspiration using the Thornthwaite equation for day \eqn{i}, and
\eqn{P_i} is the precipitation (mm) for day \eqn{i}. The constants \eqn{c_s}, \eqn{c_d}, and \eqn{c_p} were optimised based on ground truthed data from
\href{https://www.tern.org.au/news-smips-soil-moisture/}{TERN}
}
\seealso{
\code{\link[=get.BARRA.weather]{get.BARRA.weather()}}, \code{\link[=get.SILO.weather]{get.SILO.weather()}}, \code{\link[=get.CMIP6.weather]{get.CMIP6.weather()}}, \code{\link[=dl.slga]{dl.slga()}}
}
\author{
Nick Fradgley
}
