% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAIGE_data_doc.R
\docType{data}
\name{CAIGE20_22envs}
\alias{CAIGE20_22envs}
\title{CAIGE wheat trial environments data}
\format{
A data frame with 36 rows and 9 columns:
Column names include
}
\usage{
CAIGE20_22envs
}
\description{
A subset of data from the CAIGE project wheat Multi environment yield trial series in the years 2020 to 2022.
}
\keyword{datasets}
