% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAIGE data doc.R
\docType{data}
\name{CAIGE22_23envs}
\alias{CAIGE22_23envs}
\title{CAIGE wheat trial environments data}
\format{
A data frame with 36 rows and 9 columns:
Column names include
}
\usage{
CAIGE22_23envs
}
\description{
A subset of data from the CAIGE project wheat Multi environment yield trial series in the years 2020 to 2022.
}
\keyword{datasets}
