% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl.slga_function.R
\name{dl.slga}
\alias{dl.slga}
\title{Download soil data from SLGA to a directory}
\usage{
dl.slga(dir, ncores = NULL, verbose = TRUE, dlprompt = FALSE)
}
\arguments{
\item{dir}{Directory path that soil TIFF files will be saved in.}

\item{ncores}{Number (integer) of cores to use for parallel processing. Use \code{1} to run sequentially in series. The default (\code{NULL}) will
use the maximum available cores. If running in parallel, an output log text file will be created in the working directory.}

\item{verbose}{Logical. Should progress be printed? Default = TRUE.}

\item{dlprompt}{Logical. Should the user be prompted approve the total download size? Default = FALSE.}
}
\value{
No value is returned but files should be saved in the \code{dir} directory path.
}
\description{
A function to bulk download all required TIFF files for multiple soil attributes at multiple depths
from the \href{https://esoil.io/TERNLandscapes/Public/Pages/SLGA/index.html}{SLGA} data resource. TIFF files are saved in a specified directory
which can be used to point the \code{\link[=get.S.ECs]{get.S.ECs()}} to if used repeatedly. This will require a fast internet connection.
}
\references{
Grundy, M. J., Rossel, R. A. V., Searle, R. D., Wilson, P. L., Chen, C., Gregory, L. J., Grundy, M. J., Rossel,
R. A. V., Searle, R. D., Wilson, P. L., Chen, C., & Gregory, L. J. (2015).
\href{https://doi.org/10.1071/SR15191}{Soil and Landscape Grid of Australia. Soil Research}, 53(8), 835–844.
}
\seealso{
\code{\link[=get.S.ECs]{get.S.ECs()}}
}
\author{
Nick Fradgley
}
