% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.CMIP6.Weather.Function.R
\name{get.CMIP6.weather}
\alias{get.CMIP6.weather}
\title{Get weather data from BARRA-R2}
\usage{
get.CMIP6.weather(
  Envs,
  Lats,
  Lons,
  Years,
  GCMs,
  SSPs,
  ncores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{Envs}{Vector of environment names character strings.}

\item{Lats}{Vector of latitude numeric values for each environment in the same order as \code{Envs}.}

\item{Lons}{Vector of longitude numeric values for each environment in the same order as \code{Envs}.}

\item{Years}{Vector of year integer values for each environment in the same order as \code{Envs}.}

\item{GCMs}{Vector of GCM names to get data from. For options see Details below.}

\item{SSPs}{Vector of SSP names to get data from. For options see Details below.}

\item{ncores}{Number (integer) of cores to use for parallel processing of gridded data up to 5 cores. Use \code{1} to run in series. The default (\code{NULL}) will
use the maximum available cores up to 5. If running in parallel, an output log text file will be created in the working directory.}

\item{verbose}{Logical.Should progress be printed?}
}
\value{
A list of length 2:
\itemize{
\item \verb{$data} is a list of matrices of weather data for each weather variable.
Each data matrix has environment names as rows and days of the ywar as columns
\item \verb{$Env.info} is a data frame of environment names and coordinate values for environments included in the data.
}
}
\description{
Extract weather data for Coupled Model Intercomparison Project phase 6 using a Quantile Delta Change method for Australia
\href{https://doi.org/10.25919/03by-9y62}{CMIP6 QDM} weather data resource for a set of environments with defined latitude and longitude coordinates.

Weather variables include:
\itemize{
\item \code{daily_rain} - Daily rainfall (mm)
\item \code{max_temp} - Maximum temperature (°C)
\item \code{min_temp} - Minimum temperature (°C)
\item \code{vp_deficit} - Vapour pressure deficit (hPa)
\item \code{radiation} - Solar exposure, consisting of both direct and diffuse components (MJ m\if{html}{\out{<sup>}}-2\if{html}{\out{</sup>}})
}
}
\details{
The CMPI6 QDM dataset includes climate projections for two time future periods of years:
\enumerate{
\item 2035-2064
\item 2070-2099
\code{Years} values must be within either of these two time periods.
}

Possible options for Global Climate Models (GCM):
\itemize{
\item \code{ACCESS-CM2} - A much hotter future, and drier in most regions except the southeast
\item \code{ACCESS-ESM1.5} - A hotter and much drier future
\item \code{CESM2} - A hotter future, wetter in parts of the east and north
\item \code{CMCC-ESM2} - A much hotter future with little change in mean rainfall (with regional exceptions)
\item \code{CNRMESM2-1} - A much hotter future, much drier especially in the east, but wetter in the northwest
\item \code{EC-Earth3} - A hotter and much wetter future for much of Australia (except southwest Western Australia)
\item \code{MPI-ESM1-2-HR} - Lower warming, mid-range changes in rainfall
\item \code{NorESM2-MM} - Lower warming, mid-range changes in rainfall
\item \code{UKESM1-0-LL} - Low probability, high impact case (high climate sensitivity, high Australian warming)
For further details see \href{https://doi.org/10.1016/j.cliser.2023.100368}{Grose et al. 2023}
}

Possible options for Shared Socio-economic Pathways (SSP) and equivalent Representative Concentration Pathways (RCP):
\itemize{
\item \code{ssp126} - Sustainability (RCP2.6)
\item \code{ssp245} - Middle of the road (RCP4.5)
\item \code{ssp370} - Regional rivalry (No equivalent RCP)
\item \code{ssp585} - Fossil fuelled development (RCP8.5)
}

For further details see \href{https://doi.org/10.1016/j.gloenvcha.2016.05.009}{Riahi et al. 2017}
}
\references{
\itemize{
\item Grose, M. R., Narsey, S., Trancoso, R., Mackallah, C., Delage, F., Dowdy, A., Di Virgilio, G., Watterson, I., Dobrohotoff, P., Rashid, H. A., Rauniyar, S., Henley, B., Thatcher, M., Syktus, J., Abramowitz, G., Evans, J. P., Su, C.-H., & Takbash, A. (2023).
\href{https://doi.org/10.1016/j.cliser.2023.100368}{A CMIP6-based multi-model downscaling ensemble to underpin climate change services in Australia}. Climate Services, 30, 100368.
\item Irving, D., & Macadam, I. (2024). \href{https://doi.org/10.25919/03by-9y62}{Application-Ready Climate Projections from CMIP6 using the Quantile Delta Change method. CSIRO Climate Innovation Hub Technical Note 5}.
Technical Note.
\item Riahi, K., van Vuuren, D. P., Kriegler, E., Edmonds, J., O’Neill, B. C., Fujimori, S., Bauer, N., Calvin, K., Dellink, R., Fricko, O., Lutz, W., Popp, A., Cuaresma, J. C., Kc, S., Leimbach, M., Jiang, L., Kram, T., Rao, S., Emmerling, J., … Tavoni, M. (2017).
\href{https://doi.org/10.1016/j.gloenvcha.2016.05.009}{The Shared Socioeconomic Pathways and their energy, land use, and greenhouse gas emissions implications: An overview}. Global Environmental Change, 42, 153–168.
}
}
\seealso{
\code{\link[=get.SILO.weather]{get.SILO.weather()}}, \code{\link[=get.BARRA.weather]{get.BARRA.weather()}}
}
