% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict environmental effects function.R
\name{pred.env.effs}
\alias{pred.env.effs}
\title{Predict environmental effects}
\usage{
pred.env.effs(train.ECs, new.ECs, E.effs)
}
\arguments{
\item{train.ECs}{A data frame of weather and/or soil ECs for observed environments as output from the [get.W.ECs()] or `get.S.ECs()` functions.}

\item{new.ECs}{A data frame of weather and/or soil ECs for new environments as output from the `get.W.ECs()` or `get.S.ECs()` functions.}

\item{E.effs}{A data frame of several or a vector of a single Environmental effects parameters fitted from a multi-environmnet mixed model.
Latent environmental effect factor loadings that decompose GxE can be defined as described by Smith et al (2021).}
}
\value{
A data frame of environmental effect predictions for the new environments with environments as rows and environmental effect variates as columns.
}
\description{
A function to train a random forest model to predict environmental effects defined from a MET mixed model analysis with ECs as predictors.
}
\references{
Smith, A., Norman, A., Kuchel, H., & Cullis, B. (2021). Plant Variety Selection Using Interaction Classes Derived From Factor Analytic Linear Mixed Models:
     Models With Independent Variety Effects. Frontiers in Plant Science, 12. https://doi.org/10.3389/fpls.2021.737462
}
