% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_locs_function.R
\name{range.locs}
\alias{range.locs}
\title{Make a set of coordinates within the Australian cropping ranges}
\usage{
\method{range}{locs}(
  min.lon = 110,
  max.lon = 160,
  min.lat = -40,
  max.lat = -20,
  res = 0.1,
  states = c("WA", "SA", "NSW", "QLD", "VIC")
)
}
\arguments{
\item{min.lon}{Numeric. Minimum longitude value. Default value = \code{110}.}

\item{max.lon}{Numeric. Maximum longitude value. Default value = \code{160}.}

\item{min.lat}{Numeric. Minimum latitude value. Default value = \code{-40}.}

\item{max.lat}{Numeric. Maximum latitude value. Default value = \code{-20}.}

\item{res}{Numeric. Resolution, lat/lon distance between points. Default = \code{0.1}.}

\item{states}{vector of character strings. Abbreviated names of Australian states to make points for. Options include: \code{c("WA","SA","NSW","QLD","VIC","TAS","NT")}}
}
\value{
A data frame of coordinates with Loc, Lon, and Lat column names.
}
\description{
This function uses will generate a data frame of coordinates within a given boundary of lat and lon values as well
as within Australian states that are within the Australian cropping ranges according to data from the Dept of Climate Change, Energy, the Environment & Water
}
\author{
Nick Fradgley
}
