% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wheat_area_data_doc.R
\docType{data}
\name{wheat.area.envs}
\alias{wheat.area.envs}
\title{Australian wheat belt environments info}
\format{
A data frame with 5080 rows and 5 columns:
}
\usage{
wheat.area.envs
}
\description{
Lat and lon location coordinates for 5,080 environments that are within the Australian wheat belt according to
data from Monfreda et al.  and the geodata R package. Environment locations are repeated for multiple years between 2018 and 2020.
All sowing dates are for the 1st of May.
}
\keyword{datasets}
