% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.BARRA.Weather.Function.R
\name{get.BARRA.weather}
\alias{get.BARRA.weather}
\title{Get weather data from BARRA-R2}
\usage{
get.BARRA.weather(
  Envs,
  Lats,
  Lons,
  Years,
  ncores = NULL,
  verbose = TRUE,
  dlprompt = FALSE
)
}
\arguments{
\item{Envs}{Vector of environment names character strings.}

\item{Lats}{Vector of latitude numeric values for each environment in the same order as \code{Envs}.}

\item{Lons}{Vector of longitude numeric values for each environment in the same order as \code{Envs}.}

\item{Years}{Vector of year integer values for each environment in the same order as \code{Envs}.}

\item{ncores}{Number (integer) of cores to use for parallel processing of gridded data up to 5 cores. Use \code{1} to run in series. The default (\code{NULL}) will
use the maximum available cores up to 5. If running in parallel, an output log text file will be created in the working directory.}

\item{verbose}{Logical. Should progress be printed? Default if TRUE.}

\item{dlprompt}{Logical. Should the user be prompted approve the total download size? Default it TRUE.}
}
\value{
A list of length 2:
\itemize{
\item \verb{$data} is a list of matrices of weather data for each weather variable.
Each data matrix has environment names as rows and days of the ywar as columns
\item \verb{$Env.info} is a data frame of environment names and coordinate values for environments included in the data.
}
}
\description{
Extract weather data for Australia from the \href{https://opus.nci.org.au/spaces/NDP/pages/264241166/BOM+BARRA2+ob53}{BARRA-R2}
weather data resource for a set of environments with defined latitude and longitude coordinates. BARRA-R2 data runs from Jan 1979 to Sept 2024.
}
\details{
Weather variables returned include:
\itemize{
\item \code{daily_rain} - Daily rainfall (mm)
\item \code{max_temp} - Maximum temperature (°C)
\item \code{min_temp} - Minimum temperature (°C)
\item \code{vp_deficit} - Vapour pressure deficit (hPa)
\item \code{radiation} - Solar exposure, consisting of both direct and diffuse components (MJ m\if{html}{\out{<sup>}}-2\if{html}{\out{</sup>}})
\item \code{day_lengths} - Time between sunrise and sunset (h) not taken from BARRA-R2
}
}
\references{
Su, C.H., Dharssi, I., Le Marshall, J., Le, T., Rennie, S., Smith, A., Stassen, C., Steinle, P., Torrance, J., Wang, C. and Warren, R.A., 2022.
\href{http://www.bom.gov.au/research/publications/researchreports/BRR-067.pdf}{BARRA2: Development of the next-generation Australian regional atmospheric reanalysis}.
Bureau of Meteorology.
}
\seealso{
\code{\link[=get.SILO.weather]{get.SILO.weather()}}
}
