% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.CMIP6.Weather.Function.R
\name{get.CMIP6.weather}
\alias{get.CMIP6.weather}
\title{Get weather data from CMIP6 QDC}
\usage{
get.CMIP6.weather(
  Envs,
  Lats,
  Lons,
  Years,
  GCMs,
  SSPs,
  ncores = NULL,
  verbose = TRUE,
  dlprompt = FALSE
)
}
\arguments{
\item{Envs}{Vector of environment names character strings.}

\item{Lats}{Vector of latitude numeric values for each environment in the same order as \code{Envs}.}

\item{Lons}{Vector of longitude numeric values for each environment in the same order as \code{Envs}.}

\item{Years}{Vector of year integer values. Unlike \code{\link[=get.SILO.weather]{get.SILO.weather()}} and \code{\link[=get.BARRA.weather]{get.BARRA.weather()}}, \code{Years} should not be the same
length as \code{Envs}. Data for all locations in \code{Envs} will be extracted for all \code{Years}. Values must be within the possible time ranges of 1985:2014, 2035-2064,
or 2070-2099.}

\item{GCMs}{Vector of GCM names to get data from. For options see Details below.}

\item{SSPs}{Vector of SSP names to get data from. For options see Details below.}

\item{ncores}{Number (integer) of cores to use for parallel processing of gridded data up to 5 cores. Use \code{1} to run in series. The default (\code{NULL}) will
use the maximum available cores up to 5. If running in parallel, an output log text file will be created in the working directory.}

\item{verbose}{Logical. Should progress be printed? Default = TRUE.}

\item{dlprompt}{Logical. Should the user be prompted approve the total download size? Default = TRUE.}
}
\value{
A multi-level list of \verb{$data} and \verb{$Env.info} weather variables within SSPs within GCMs:
\itemize{
\item \verb{$data} is a list of matrices of weather data for each weather variable.
Each data matrix has environment names as rows and days of the year as columns
\item \verb{$Env.info} is a data frame of environment names and coordinate values for environments included in the data.
}

E.g. \code{x$ACCESS-CM2$ssp585$daily_rain$data[1,1]} will return the rainfall on the Jan 1st at the first location in the
first year for the high emissions SSP585 for the ACCESS-CM2 GCM.
}
\description{
Extract weather data from the Coupled Model Intercomparison Project phase 6 using a Quantile Delta Change method for Australia
\href{https://doi.org/10.25919/03by-9y62}{CMIP6 QDC} weather data resource for a set of environments with defined latitude and longitude coordinates.
}
\details{
The CMPI6 QDC dataset is hosted on the \href{https://data-cbr.csiro.au/thredds/catalog/catch_all/qdc-cmip6/QDC-CMIP6/BARRA-R2/catalog.html}{CSIRO data server}
includes climate projections for historical (1985-2014) and two future periods of years (2035-2064 and 2070-2099). All \code{Years} values must be within these time periods.
The QDC method benchmarks CMIP6 against the observational data from the \href{https://opus.nci.org.au/spaces/NDP/pages/264241166/BOM+BARRA2+ob53}{BARRA-R2}
dataset so should be used in combination with the \code{\link[=get.BARRA.weather]{get.BARRA.weather()}} rather than the \code{\link[=get.SILO.weather]{get.SILO.weather()}} function for observed environments that are
outside of the 1985-2014 time period of the CMIP6 QDC historical time period.

Weather variables returned include:
\itemize{
\item \code{daily_rain} - Daily rainfall (mm)
\item \code{max_temp} - Maximum temperature (°C)
\item \code{min_temp} - Minimum temperature (°C)
\item \code{vp_deficit} - Vapour pressure deficit (hPa)
\item \code{radiation} - Solar exposure, consisting of both direct and diffuse components (MJ m\if{html}{\out{<sup>}}-2\if{html}{\out{</sup>}})
\item \code{day_lengths} - Time between sunrise and sunset (h) not taken from CMIP6 QDC
}

VPD in hPa is calculated as \eqn{ VPD = 10(es - ea) }, where \eqn{ es = 0.6108 \times \exp(\frac{17.27 \times T_{ave}}{T_{ave} + 237.3}) },
\eqn{T_{ave} } is the mean temperature in °C, \deqn{ ea = \frac{RH}{100} \times es }, and \eqn{RH} is the relative humidity (\%).

Possible options for Global Climate Models (GCM):
\itemize{
\item \code{ACCESS-CM2} - A much hotter future, and drier in most regions except the southeast
\item \code{ACCESS-ESM1-5} - A hotter and much drier future
\item \code{CMCC-ESM2} - A much hotter future with little change in mean rainfall (with regional exceptions)
\item \code{CNRM-ESM2-1} - A much hotter future, much drier especially in the east, but wetter in the northwest
\item \code{EC-Earth3} - A hotter and much wetter future for much of Australia (except southwest Western Australia)
\item \code{MPI-ESM1-2-HR} - Lower warming, mid-range changes in rainfall
\item \code{NorESM2-MM} - Lower warming, mid-range changes in rainfall
\item \code{UKESM1-0-LL} - Low probability, high impact case (high climate sensitivity, high Australian warming)
}

For further details see \href{https://doi.org/10.1016/j.cliser.2023.100368}{Grose et al. 2023}

Possible options for Shared Socio-economic Pathways (SSP) and equivalent Representative Concentration Pathways (RCP)
with expected temperature increase range:
\itemize{
\item \code{ssp126} - Sustainability (RCP2.6; 1.0-1.8°C)
\item \code{ssp245} - Middle of the road (RCP4.5; 1.3-2.4°C)
\item \code{ssp370} - Regional rivalry (No equivalent RCP; 2.8-4.6°C)
\item \code{ssp585} - Fossil fueled development (RCP8.5; 3.3-5.7°C)
}

For further details see \href{https://doi.org/10.1016/j.gloenvcha.2016.05.009}{Riahi et al. 2017}

An internet connection with high download speed is suggested for downloading gridded data for many environments.
}
\references{
\itemize{
\item Grose, M. R., Narsey, S., Trancoso, R., Mackallah, C., Delage, F., Dowdy, A., Di Virgilio, G., Watterson, I., Dobrohotoff, P., Rashid, H. A., Rauniyar, S., Henley, B., Thatcher, M., Syktus, J., Abramowitz, G., Evans, J. P., Su, C.-H., & Takbash, A. (2023).
\href{https://doi.org/10.1016/j.cliser.2023.100368}{A CMIP6-based multi-model downscaling ensemble to underpin climate change services in Australia}. Climate Services, 30, 100368.
\item Irving, D., & Macadam, I. (2024). \href{https://doi.org/10.25919/03by-9y62}{Application-Ready Climate Projections from CMIP6 using the Quantile Delta Change method. CSIRO Climate Innovation Hub Technical Note 5}.
Technical Note.
\item Riahi, K., van Vuuren, D. P., Kriegler, E., Edmonds, J., O’Neill, B. C., Fujimori, S., Bauer, N., Calvin, K., Dellink, R., Fricko, O., Lutz, W., Popp, A., Cuaresma, J. C., Kc, S., Leimbach, M., Jiang, L., Kram, T., Rao, S., Emmerling, J., … Tavoni, M. (2017).
\href{https://doi.org/10.1016/j.gloenvcha.2016.05.009}{The Shared Socioeconomic Pathways and their energy, land use, and greenhouse gas emissions implications: An overview}. Global Environmental Change, 42, 153–168.
}
}
\seealso{
\code{\link[=get.SILO.weather]{get.SILO.weather()}}, \code{\link[=get.BARRA.weather]{get.BARRA.weather()}}
}
\author{
Nick Fradgley
}
