% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.Soil.EC_function.R
\name{get.S.ECs}
\alias{get.S.ECs}
\title{Get environmental covariates from soils data}
\usage{
get.S.ECs(Envs, Lats, Lons, ncores = NULL, verbose = TRUE, dlprompt = FALSE)
}
\arguments{
\item{Envs}{Vector of environment names character strings.}

\item{Lats}{Vector of latitude numeric values for each environment.}

\item{Lons}{Vector of longitude numeric values for each environment.}

\item{ncores}{Number (integer) of cores to use for parallel processing. Use \code{1} to run sequentially in series. The default (\code{NULL}) will
use the maximum available cores. If running in parallel, an output log text file will be created in the working directory.}

\item{verbose}{Logical. Should progress be printed?}

\item{dlprompt}{Logical. Should the user be prompted approve the total download size? Default it FALSE.}
}
\value{
A data frame of soil EC values with environment names as rows and covariates as columns.
}
\description{
A function to derive soil Environmental Covariates (ECs) for multiple soil attributes at multiple depths
from the \href{https://esoil.io/TERNLandscapes/Public/Pages/SLGA/index.html}{SLGA} data resource.
}
\references{
Grundy, M. J., Rossel, R. A. V., Searle, R. D., Wilson, P. L., Chen, C., Gregory, L. J., Grundy, M. J., Rossel,
R. A. V., Searle, R. D., Wilson, P. L., Chen, C., & Gregory, L. J. (2015).
\href{https://doi.org/10.1071/SR15191}{Soil and Landscape Grid of Australia. Soil Research}, 53(8), 835–844.
}
\seealso{
\code{\link[=get.W.ECs]{get.W.ECs()}}
}
\author{
Nick Fradgley
}
