% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.Soil.EC function.R
\name{get.S.ECs}
\alias{get.S.ECs}
\title{Get environmental covariates from soils data}
\usage{
get.S.ECs(Envs, Lats, Lons, verbose = TRUE)
}
\arguments{
\item{Envs}{Vector of environment names character strings.}

\item{Lats}{Vector of latitude numeric values for each environment.}

\item{Lons}{Vector of longitude numeric values for each environment.}

\item{verbose}{Logical. Should progress be printed?}
}
\value{
A data frame of soil EC values with environment names as rows and covariates as columns.
}
\description{
A function to derive soil Environmental Covariates (ECs) from the Soils and Landscapes Grids Australia data resource.
https://esoil.io/TERNLandscapes/Public/Pages/SLGA/index.html
}
\references{
Grundy, M. J., Rossel, R. A. V., Searle, R. D., Wilson, P. L., Chen, C., Gregory, L. J., Grundy, M. J., Rossel,
     R. A. V., Searle, R. D., Wilson, P. L., Chen, C., & Gregory, L. J. (2015).
     Soil and Landscape Grid of Australia. Soil Research, 53(8), 835–844.
     https://doi.org/10.1071/SR15191
}
