% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.SILO.Weather.Function.R
\name{get.SILO.weather}
\alias{get.SILO.weather}
\title{Get weather data from SILO}
\usage{
get.SILO.weather(Envs, Lats, Lons, Years, verbose = TRUE)
}
\arguments{
\item{Envs}{Vector of environment names character strings.}

\item{Lats}{Vector of latitude numeric values for each environment in the same order as \code{Envs}.}

\item{Lons}{Vector of longitude numeric values for each environment in the same order as \code{Envs}.}

\item{Years}{Vector of year integer values for each environment in the same order as \code{Envs}.}

\item{verbose}{Logical. Should progress be printed?}
}
\value{
A list of weather data for each weather variable and a vector of units for each variable.
The data list contains a a matrix or data values with environments as rows and days of the year as columns.
}
\description{
Extract weather data for Australia from the \href{https://www.longpaddock.qld.gov.au/silo/}{SILO} weather data resource
for a set of environments with defined latitude and longitude coordinates.

Weather variables include:
\itemize{
\item \code{daily_rain} - Daily rainfall (mm)
\item \code{max_temp} - Maximum temperature (deg C)
\item \code{min_temp} - Minimum temperature (deg C)
\item \code{vp_deficit} - Vapour pressure deficit (hPa)
\item \code{radiation} - Solar exposure, consisting of both direct and diffuse components (MJ/m2)
}
}
\details{
When there are only a few environments, point data will be sequentially downloaded from SILO. When there are many environments in each year,
data will be downloaded and extracted from whole gridded data files more efficiently.
}
\references{
Jeffrey, S. J., Carter, J. O., Moodie, K. B., & Beswick, A. R. (2001).
Using spatial interpolation to construct a comprehensive archive of Australian climate data. Environmental Modelling & Software, 16(4), 309–330.
\url{https://doi.org/10.1016/S1364-8152(01)00008-1}
}
