% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.SILO.Weather.Function.R
\name{get.SILO.weather}
\alias{get.SILO.weather}
\title{Get weather data from SILO}
\usage{
get.SILO.weather(Envs, Lats, Lons, Years, ncores = NULL, verbose = TRUE)
}
\arguments{
\item{Envs}{Vector of environment names character strings.}

\item{Lats}{Vector of latitude numeric values for each environment in the same order as \code{Envs}.}

\item{Lons}{Vector of longitude numeric values for each environment in the same order as \code{Envs}.}

\item{Years}{Vector of year integer values for each environment in the same order as \code{Envs}.}

\item{ncores}{Number (integer) of cores to use for parallel processing of gridded data up to 5 cores. Use \code{1} to run in series. The default (\code{NULL}) will
use the maximum available cores up to 5. If running in parallel, an output log text file will be created in the working directory.}

\item{verbose}{Logical. Should progress be printed?}
}
\value{
A list of length 2:
\itemize{
\item \verb{$data} is a list of matrices of weather data for each weather variable.
Each data matrix has environment names as rows and days of the year as columns
\item \verb{$Env.info} is a data frame of environment names and coordinate values for environments included in the data.
}
}
\description{
Extract weather data for Australia from the \href{https://www.longpaddock.qld.gov.au/silo/}{SILO} weather data resource
for a set of environments with defined latitude and longitude coordinates.

Weather variables include:
\itemize{
\item \code{daily_rain} - Daily rainfall (mm)
\item \code{max_temp} - Maximum temperature (°C)
\item \code{min_temp} - Minimum temperature (°C)
\item \code{vp_deficit} - Vapour pressure deficit (hPa)
\item \code{radiation} - Solar exposure, consisting of both direct and diffuse components (MJ m\if{html}{\out{<sup>}}-2\if{html}{\out{</sup>}})
}
}
\details{
When there are only a few environments, point data will be sequentially downloaded from SILO. When there are many environments in each year,
data will be downloaded and extracted from whole gridded data files more efficiently. Any locations outside of the Australian land area will return \code{NA}.
}
\references{
Jeffrey, S. J., Carter, J. O., Moodie, K. B., & Beswick, A. R. (2001).
\href{https://doi.org/10.1016/S1364-8152(01)00008-1}{Using spatial interpolation to construct a comprehensive archive of Australian climate data. Environmental Modelling & Software},
16(4), 309–330.
}
\seealso{
\code{\link[=get.BARRA.weather]{get.BARRA.weather()}}
}
