% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.SILO.Weather.Function.R
\name{Get.SILO.weather}
\alias{Get.SILO.weather}
\title{Get weather data from SILO}
\usage{
Get.SILO.weather(
  Envs,
  Lats,
  Lons,
  Years,
  vars = c("daily_rain", "max_temp", "min_temp", "vp_deficit", "radiation")
)
}
\arguments{
\item{Envs}{Vector of environment names character strings.}

\item{Lats}{Vector of latitude numeric values for each environment.}

\item{Lons}{Vector of longitude numeric values for each environment.}

\item{Years}{Vector of year integer values for each environment.}

\item{vars}{Vector of weather variable names to get. Default is all.
Options include:

   *daily_rain - Daily rainfall (mm)

   *max_temp - Maximum temperature (deg C)

   *min_temp - Minimum temperature (deg C)

   *vp_deficit - Vapour pressure deficit (hPa)

   *radiation - Solar exposure, consisting of both direct and diffuse components (MJ/m2)}
}
\value{
A list of weather data for each weather variable and a vector of units for each variable.
The data list contains a a matrix or data values with environments as rows and days of the year as columns.
}
\description{
Extract weather data from the SILO weather data resource (https://www.longpaddock.qld.gov.au/silo/)
for a set of environments with defined latitude and longitude coordinates
}
\examples{
data("CAIGE23_24envs")
wthr <- Get.SILO.weather()

}
\references{
Jeffrey, S. J., Carter, J. O., Moodie, K. B., & Beswick, A. R. (2001).
  Using spatial interpolation to construct a comprehensive archive of Australian climate data. Environmental Modelling & Software, 16(4), 309–330.
  https://doi.org/10.1016/S1364-8152(01)00008-1
}
